Break out resource tests into separate files.
