+++ stripe/test/__init__.py
@@ -1,3 +1,4 @@
+import os
 import pkgutil
 import unittest2
 
@@ -5,11 +6,12 @@ import unittest2
 def all_names():
     for _, modname, _ in pkgutil.iter_modules(__path__):
         if modname.startswith('test_'):
-            yield 'stripe.test.' + modname
+            yield 'tests.' + modname
 
 
 def all():
-    return unittest2.defaultTestLoader.loadTestsFromNames(all_names())
+    path = os.path.dirname(os.path.realpath(__file__))
+    return unittest2.defaultTestLoader.discover(path)
 
 
 def unit():
